requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# ML Libraries
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Kafka
kafka-python>=2.0.2

# Web framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# HTTP client
requests>=2.31.0

# Optional: Streamlit dashboard
# streamlit>=1.25.0

# Development
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
